add loading to getTracks functionality
