+++ client/src/components/Home.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { styled } from '@mui/system';
 import {
   Box,
@@ -7,10 +7,11 @@ import {
   Button,
   TextField,
   GlobalStyles,
+  CircularProgress
 } from '@mui/material';
 import heroImage from '../../public/assets/heroImage.png';
 import GoogleOauth from './GoogleOauth';
-import { spotifyThemeColor, youtTubeThemeColor } from '../theme/customTheme';
+import customTheme, { spotifyThemeColor, youtTubeThemeColor } from '../theme/customTheme';
 
 const HeroContent = styled(Container)(({ theme }) => ({
   display: 'flex',
@@ -133,7 +134,41 @@ const HeroImage = styled(Box)(({ theme }) => ({
   // border: '2px solid green',
 }));
 
-function Home({ getSpotifyApi, getTracks, googleOauth }) {
+function Home({ getSpotifyApi, googleOauth }) {
+  const [loading, setLoading] = useState(false);
+
+  // functionality
+  const getTracks = ()=>{
+    console.log('clicked button');
+    const textBox = document.getElementById('spotify-playlist');
+    let playlist_id = textBox.value.split('/');
+    playlist_id = playlist_id[playlist_id.length - 1];
+    playlist_id = playlist_id.split('?')[0];
+    console.log('PLAYLIST ID: ', playlist_id);
+
+    // show loading
+    setLoading(true);
+
+    fetch('http://localhost:3000/spotify/getTrackData', {
+      method:'POST',
+      headers:{'Content-Type': 'application/json'},
+      body: JSON.stringify({token: window.localStorage.getItem('spotify_access_token'),
+            playlist_id: playlist_id
+      })})
+      .then(res=>{
+        // hide loading
+        setLoading(false);
+        console.log('res in get trackers', res);
+        res.json().then(playlist_info=>{
+          console.log('playlist_info in gettracker', playlist_info);
+        });
+      }).catch(err=>{
+        // hide loading
+        setLoading(false);
+        console.log(err);
+      })
+    }
+
   return (
     <>
       <GlobalStyles
@@ -159,7 +194,9 @@ function Home({ getSpotifyApi, getTracks, googleOauth }) {
           placeholder='Enter Spotify URL'
         />
         <TransferNowButton onClick={getTracks}>Transfer Now</TransferNowButton>
+        {loading && <CircularProgress color='primary'/>}
       </HeroContent>
+
       <HeroImage className='HeroImage' />
     </>
   );
