+++ client/src/containers/HomeContainer.jsx
@@ -46,30 +46,6 @@ const getSpotifyApi = ()=>{
 
 }
 
-const getTracks = ()=>{
-  console.log('clicked button');
-  const textBox = document.getElementById('spotify-playlist');
-  let playlist_id = textBox.value.split('/');
-  playlist_id = playlist_id[playlist_id.length - 1];
-  playlist_id = playlist_id.split('?')[0];
-  console.log('PLAYLIST ID: ', playlist_id);
-
-    fetch('http://localhost:3000/spotify/getTrackData', {
-      method:'POST',
-      headers:{'Content-Type': 'application/json'},
-      body: JSON.stringify({token: window.localStorage.getItem('spotify_access_token'),
-            playlist_id: playlist_id
-      })})
-      .then(res=>{
-        console.log('res in get trackers', res);
-        res.json().then(playlist_info=>{
-          console.log('playlist_info in gettracker', playlist_info);
-        });
-      }).catch(err=>{
-        console.log(err);
-      })
-    }
-
   const googleOauth = async () => {
     try {
       const response = await fetch('api/google');
@@ -96,7 +72,7 @@ function HomeContainer() {
 
   return (
     <FlexContainer className='FlexContainer' maxWidth='xl'>
-        <Home getSpotifyApi={getSpotifyApi} getTracks={getTracks} googleOauth={googleOauth}/>
+        <Home getSpotifyApi={getSpotifyApi} googleOauth={googleOauth}/>
     </FlexContainer>
   )
 }
